+++ bauh/gems/arch/pacman.py
@@ -514,7 +514,7 @@ def get_update_size(pkgs: List[str]) -> Dict[str, int]:  # bytes:
     return {}
 
 
-def get_installed_size(pkgs: List[str]) -> Dict[str, int]: # bytes
+def get_installed_size(pkgs: List[str]) -> Dict[str, int]:  # bytes
     output = run_cmd('pacman -Qi {}'.format(' '.join(pkgs)))
 
     if output: