+++ CHANGELOG.md
@@ -49,6 +49,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
         - faster for scenarios involving several packages ( taking =~ 95% less time )
         - faster for AUR installations ( taking an average of 23% less time )
     - code refactored
+    - displaying missing repository dependencies sizes
     - overall speed improvements 
 - UI
     - table update performance